add M2

+++ M2 Zahlen/Double.java
@@ -3,7 +3,7 @@ import java.util.concurrent.TimeUnit;
 public class Double {
     public static void main(String[] args) throws InterruptedException {
         // Problem 1
-        double d = (double) 1 / 2;
+        double d = 1 / 2;
         System.out.println(d);
 
         // Problem 2
